cmake: Replaced CMAKE_CURRENT_SOURCE_DIR with CMAKE_SOURCE_DIR
cr_add_subproejct didn't work correctly when Criterion was used as a git
submoudle in another project as follows.

    $ cd my_git_project
    $ mkdir vendor
    $ git submodule add \
        https://github.com/Snaipe/Criterion.git vendor/criterion

The cause of the problem was the use of CMAKE_SOURCE_DIR.
CMAKE_SOURCE_DIR was expanded to the path to the top level of the source
tree.

The problem was fixed by replacing CMAKE_SOURCE_DIR with
CMAKE_CURRENT_SOURCE_DIR which was expanded to the path to the source
directory currently being processed.