setjmp: Don't use the windows CRT implementation for setjmp/longjmp
The default CRT implementation of setjmp and longjmp unwinds the stack.

This breaks C programs expecting otherwise -- with BoxFort, the stack
unwinding gets broken due to the absolute jmp at the start of the main.
This makes everything more complicated w.r.t. actually long-jumping and
just fails completely when calling setjmp in succession on the same
jmp_buf. Go figure.

So now, we have to add some special handling for the special snowflake
that is Windows: use __builtin_setjmp/longjmp on MinGW-w64, and
our own implementation that resides in BoxFort.

Hopefully this doesn't break in the future.

Fixes #150.