stats: Stop reporting passing assertions.
Before this fix, assertions were always reported to the runner by the
test, even when passing, for accuracy w.r.t statistics.

This wasn't really a problem as prior to 2.3.x, criterion was using
a very simple protocol to pass its data, which could handle passing
extra data for correctness.

However, since the I/O layer rewrite that introduced a more complex protocol
format between the runner and the tests, passing repeatedly extra data
for passed assertions incurred a huge slowdown for large number of
assertions.

In this situation, and because data from passed assertions is rarely
used, we provide a mean to trade performance at the cost of
correctness: We introduce a `full_stats` mode, which when activated
reproduces the old behaviour, and a new behaviour in which tests stop
reporting passing assertions, and only send an update to the count of
passing assertions (without any detail attached).

Users implementing custom reporting mechanisms in which having full
access to the data of passing assertions will still be able to access it
by using the `--full-stats` CLI switch.

This fixes #172.