samples: updated parameterized tests to use the criterion-specific allocator
The current sample confused people, as it wasn't clear why using strings
as test parameters would suddenly make the whole test crash.

This was because, since the strings were read-only constants, and
criterion has no way of knowing what the underlying type of the
parameter is, the rodata address wouldn't get translated into the child
process, hence the usage of cr_malloc to make this work.

The updated samples properly reflect that with the usage of a cr_strdup
function and std::basic_string with criterion::allocator<char>.