cli: add --color and --encoding.
We introduce --color=<always|never|auto> for consistency with the rest
of the world that follow this convention to control coloring. TERM=dumb,
in this context, still means --color=never unless overriden.

Similarly, we provide a --encoding switch to set the preferred output
encoding. By default, this encoding is automatically deduced from the
locale.

--ascii and CRITERION_USE_ASCII are now deprecated in favor of the above
methods.

Fixes #200.