Make cr_file_match_str fail if str is longer.
When using cr_file_match_str, if the provided string we want to match
the file against is longer than the contents of the file, the function
returns that the file matches the string.

This function is used by many other assertion routines (namely those
asserting the contents of the standard output streams). This can be very
error-prone when designing tests such as the following:

Test(criterion, cr_file_match_str, .init = cr_redirect_stdout)
{
  printf("%s%d%s", "string", 42, "other");
  fflush(stdout);
  cr_assert_stdout_eq_str("string42othernonsense");
}

This commit makes it so that cr_file_match_str detects this kind of
mismatch.

Note: It does not alter the fact that the function consumes all of the
stream.

Fix unit tests. `std::cin >> s` reads only the first word from the
stream, while we want the entire input line to be read.