cmake,deps: Fixed dependencies installing when not needed
Previously, the build system would erroneously install
BoxFort to $DESTDIR at the full binary path -- this caused
problems for package managers, where a make install DESTDIR=...
would populate the tree with boxfort files for $(pwd),
which would cause conflicts when installing them (plus it
installed things outside of the prefix, which is bad).

This is a quick & dirty fix for that. I'm not too convinced
about using ExternalProject_Add anymore for dependencies.