`cr_file_match_str` succeeds on empty file&string.
The following test would not run correctly:

```c
Test(putstr_second, test_outputs, .init = cr_redirect_stdout) {
  cr_assert_stdout_eq_str("");
}
```

This is due to the `matches` variable being set to 0 by default. When
the file is empty, the main loop is not run and `matches` never gets the
chance to be updated. The function thus returned that the file and the
string do not match.

A unit tests has been added to prevent further regression.

An `if` statement has also been added before the consuming loop in
`cr_file_match_str` which avoids an unnecessary call to `fread` if we
already know that the stream is empty.