deps: updated boxfort to 106d7da4b365e8a775a6d67c423fb85777d8d81e
This fixes issues with msync on WSL, huge slowdows for
systems supporting a large number of file descriptors, and
the stack being marked executable.

This also adds support for ARM and ARM64. This support is not CI-backed,
so it might break at any point, but it's better than letting the code
rot since people are starting to have some interest in running criterion
on ARM.

fixes #248
fixes #233
fixes #224