api: use the correct printf specifiers on mingw-gcc.
MinGW uses by default the windows CRT for printf, but can be convinced
to use its own ANSI-compatible printf implementation by defining the
__USE_MINGW_ANSI_STDIO macro.

This fix is two-fold: first, we use the "I64" specifier prefix instead of "ll"
for the long long type modifier if we're using MSVC.

Then, since cr_asprintf uses printf and is compiled in libcriterion.so,
we have to make a choice on which printf to use for our users.
Naturally, we prefer the standards-compliant one.