ci: automate release
When a commit is pushed to the master branch or a version tag is created,
the release workflow starts automatically.

The release workflow creates the source tarball (which includes
all subprojects), builds Criterion from the tarball, and creates binary
packages.

Binary packages are created within a CentOS 7 container. This sounds
terrible, but it ensures that Criterion is built with an older glibc
version (v2.17) so that it will work on a wide variety of Linux distros.

The source tarball and the binaries then will be attached to the release
on the Releases page.

In case new commits are pushed to the master branch, a pre-release will be
published automatically using git describe --long for the name of the
pre-release (this creates unsigned lightweight tags).

When the workflow is triggered by a new version tag (vX.Y.Z), a final
release DRAFT will be made. This needs to be reviewed and published
manually.

The workflow verifies the release tag as well, it opens an issue to warn
maintainers about unsigned or lightweight release tags.