diff: fix crash in assert failure reporting when there is no diff
If the user-provided function to stringify an object is bugged and
produces equal output for non-equal objects, thus producing an empty
diff, Criterion would crash trying to access an unset parameter list.

This fixes the crash by checking that we do have expected values, and
printing placeholder messages if we do not.